requirements.txt
# --- BACKEND ML STACK (Qwen3-VL Local Support) ---
# NOTE: For CUDA 12.4 support on Windows, run the following command manually:
# pip install torch==2.6.0 torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu124


# Web Framework
Flask==3.1.2
pyngrok>=7.1.0

# Configuration
PyYAML==6.0.3
python-dotenv==1.2.1
huggingface-hub==0.36.0

# API Clients (Google Gemini)
google-genai==1.57.0

# System Monitoring
psutil==6.1.1
GPUtil==1.4.0

# Computer Vision (Shot Detection)
scenedetect[opencv]==0.6.7.1
opencv-python==4.11.0.86
numpy==2.3.5

# Machine Learning (Local Qwen3-VL Inference)
#torch==2.6.0
#torchvision==0.21.0
transformers==4.57.3
peft==0.18.1
accelerate==1.12.0
bitsandbytes==0.49.1
pillow==12.0.0

# Serialization
orjson==3.11.5